+++ src/main/java/com/rezztoran/rezztoranbe/service/AuthService.java
@@ -2,6 +2,7 @@ package com.rezztoran.rezztoranbe.service;
 
 import com.rezztoran.rezztoranbe.dto.TokenDTO;
 import com.rezztoran.rezztoranbe.dto.request.LoginModel;
+import com.rezztoran.rezztoranbe.dto.request.PasswordResetModel;
 import com.rezztoran.rezztoranbe.dto.request.RegisterModel;
 import com.rezztoran.rezztoranbe.model.User;
 
@@ -14,4 +15,8 @@ public interface AuthService {
   boolean checkForPermission(Long id);
 
   User tryRegister(RegisterModel registerModel);
+
+  void resetPasswordRequestCodeGenerate(String email);
+
+  void resetPassword(PasswordResetModel passwordResetModel);
 }
